requirements.txt
httpx>=0.27
pydantic>=2.7
chromadb>=0.5
python-dotenv>=1.0
tiktoken>=0.7
numpy>=1.26
faiss-cpu>=1.7
sentence-transformers>=2.7
